+++ src/jvm/main/org/jetbrains/kotlinx/lincheck/strategy/managed/ManagedStrategy.kt
@@ -174,6 +174,16 @@ abstract class ManagedStrategy(
             useClocks = UseClocks.ALWAYS
         )
 
+    /**
+     * Adds a new [obj] to the snapshot, which fields will be tracked and restored during execution.
+     * This method is not supposed to be called during execution, its purpose is to account for general purpose model checker functionality.
+     * @see [runConcurrentTest].
+     */
+    fun updateSnapshotWithNewRoot(obj: Any?) {
+        staticMemorySnapshot.trackRoot(obj)
+    }
+
+
     // == STRATEGY INTERFACE METHODS ==
 
     /**
@@ -1117,7 +1127,7 @@ abstract class ManagedStrategy(
      * Required as a trick to overcome issue with leaking this in constructors, see https://github.com/JetBrains/lincheck/issues/424.
      */
     override fun updateSnapshotBeforeConstructorCall(objs: Array<Any?>) = runInIgnoredSection {
-        staticMemorySnapshot.trackObjects(objs)
+        staticMemorySnapshot.trackObjectsEagerly(objs)
     }
 
     /**
