+++ src/jvm/test/resources/expected_logs/atomic_refs_two_fields_trace_run_with_lambda.txt
@@ -1,38 +1,38 @@
 = Invalid execution results =
-| ------------------------------------------------ |
-|                     Thread 1                     |
-| ------------------------------------------------ |
-| run(() -> kotlin.Unit): IllegalStateException #1 |
-| ------------------------------------------------ |
+| -------------------------------------- |
+|                Thread 1                |
+| -------------------------------------- |
+| run(() -> R): IllegalStateException #1 |
+| -------------------------------------- |
 
 ---
 The number next to an exception name helps you find its stack trace provided after the interleaving section
 ---
 
 The following interleaving leads to the error:
-| ------------------------------------------------ |
-|                     Thread 1                     |
-| ------------------------------------------------ |
-| run(() -> kotlin.Unit): IllegalStateException #1 |
-| ------------------------------------------------ |
+| -------------------------------------- |
+|                Thread 1                |
+| -------------------------------------- |
+| run(() -> R): IllegalStateException #1 |
+| -------------------------------------- |
 
 Exception stack traces:
 #1: java.lang.IllegalStateException: Check failed.
 	at org.jetbrains.kotlinx.lincheck_test.representation.AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(Unknown Source)
-	at org.jetbrains.kotlinx.lincheck_test.representation.BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(Unknown Source)
+	at org.jetbrains.kotlinx.lincheck_test.representation.AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(Unknown Source)
 	at org.jetbrains.kotlinx.lincheck_test.representation.BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(Unknown Source)
 	at java.base/java.lang.Thread.run(Thread.java:833)
 
 Detailed trace:
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-|                                                                                                        Thread 1                                                                                                        |
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| run(() -> kotlin.Unit): IllegalStateException #1                                                                                                                                                                       |
-|   failure$1.invoke(): threw IllegalStateException at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:34)                                              |
-|     AtomicReferencesFromMultipleFieldsRunWithLambdaTest#1.block(): threw IllegalStateException at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:35) |
-|       atomicReference1.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:137)                                                                   |
-|       atomicReference2.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:137)                                                                   |
-|       AtomicReference#1.get(): Node#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:137)                                                                            |
-|       AtomicReference#1.compareAndSet(Node#1,Node#2): true at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:137)                                                       |
-|   result: IllegalStateException #1                                                                                                                                                                                     |
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+|                                                                                                       Thread 1                                                                                                       |
+| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| run(() -> R): IllegalStateException #1                                                                                                                                                                               |
+|   AtomicReferencesFromMultipleFieldsRunWithLambdaTest#1.block(): threw IllegalStateException at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:41) |
+|     block(): threw IllegalStateException at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:131)                                                                       |
+|       atomicReference1.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:143)                                                                 |
+|       atomicReference2.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:143)                                                                 |
+|       AtomicReference#1.get(): Node#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:143)                                                                          |
+|       AtomicReference#1.compareAndSet(Node#1,Node#2): true at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:143)                                                     |
+|   result: IllegalStateException #1                                                                                                                                                                                   |
+| -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
